confiratie pop-up verwijdern/wijzigen
er staan nu confirmatie pop-up voor het verwijderen en wijzigen van apparaaten

+++ admin_php/functies/InventarisVFunctie.php
@@ -19,41 +19,47 @@ echo $beschrijving;
 echo $image;
 echo $link;
 
-// Delete EXEMPLAAR_ITEM rows
-$deleteExemplaarItemQuery = "DELETE FROM EXEMPLAAR_ITEM WHERE item_id='$item_id'";
+if(isset($_POST['submitForm'])){
 
-if ($conn->query($deleteExemplaarItemQuery) === TRUE) {
-    echo "Record deleted successfully";
-} else {
-    echo "Error deleting record: " . $conn->error;
-}
+    // Delete EXEMPLAAR_ITEM rows
+    $deleteExemplaarItemQuery = "DELETE FROM EXEMPLAAR_ITEM WHERE item_id='$item_id'";
 
-// Delete FUNCTIONALITEIT rows
-$deleteFunctionaliteitQuery = "DELETE FROM FUNCTIONALITEIT WHERE item_id='$item_id'";
-if ($conn->query($deleteFunctionaliteitQuery) === TRUE) {
-    echo "Record deleted successfully";
-} else {
-    echo "Error deleting record: " . $conn->error;
-}
+    if ($conn->query($deleteExemplaarItemQuery) === TRUE) {
+        echo "Record deleted successfully";
+    } else {
+        echo "Error deleting record: " . $conn->error;
+    }
 
-// Get image_id from ITEM table
-$imageIdQuery = "SELECT image_id FROM ITEM WHERE item_id='$item_id'";
-$result = $conn->query($imageIdQuery);
-$row = $result->fetch_assoc();
-$image_id = $row['image_id'];
+    // Delete FUNCTIONALITEIT rows
+    $deleteFunctionaliteitQuery = "DELETE FROM FUNCTIONALITEIT WHERE item_id='$item_id'";
+    if ($conn->query($deleteFunctionaliteitQuery) === TRUE) {
+        echo "Record deleted successfully";
+    } else {
+        echo "Error deleting record: " . $conn->error;
+    }
 
-// Delete ITEM row
-$deleteItemQuery = "DELETE FROM ITEM WHERE item_id='$item_id'";
-if ($conn->query($deleteItemQuery) === TRUE) {
-    echo "Record deleted successfully";
-} else {
-    echo "Error deleting record: " . $conn->error;
-}
+    // Get image_id from ITEM table
+    $imageIdQuery = "SELECT image_id FROM ITEM WHERE item_id='$item_id'";
+    $result = $conn->query($imageIdQuery);
+    $row = $result->fetch_assoc();
+    $image_id = $row['image_id'];
 
-// Delete IMAGE row
-$deleteImageQuery = "DELETE FROM Images WHERE image_id='$image_id'";
-if ($conn->query($deleteImageQuery) === TRUE) {
-    echo "Record deleted successfully";
-} else {
-    echo "Error deleting record: " . $conn->error;
+    // Delete ITEM row
+    $deleteItemQuery = "DELETE FROM ITEM WHERE item_id='$item_id'";
+    if ($conn->query($deleteItemQuery) === TRUE) {
+        echo "Record deleted successfully";
+    } else {
+        echo "Error deleting record: " . $conn->error;
+    }
+
+    // Delete IMAGE row
+    $deleteImageQuery = "DELETE FROM Images WHERE image_id='$image_id'";
+    if ($conn->query($deleteImageQuery) === TRUE) {
+        echo "Record deleted successfully";
+    } else {
+        echo "Error deleting record: " . $conn->error;
+    }
+    $conn->close();
+    header('Location: ../Inventaris.php');
+
 }
